+++ CHANGELOG.md
@@ -3,7 +3,9 @@
 ### 2.14.0
 * passing the `$compression` argument as a string or `null` to `BackupExport::compression()` had been deprecated and has
   been removed (backwards compatibility removed);
-* `DeleteAllCommand`, `RotateCommand` and `SendCommand` classes had been deprecated and have been removed.
+* `DeleteAllCommand`, `RotateCommand` and `SendCommand` classes had been deprecated and have been removed;
+* `getAbsolutePath()`, `getCompression()`, `getExtension()` and `getValidCompressions()` methods provided by 
+  `BackupTrait` had been deprecated and has been removed.
 
 ## 2.13 branch
 ### 2.13.5